struct: Move all graph things to one folder
