+++ app/src/main/kotlin/app/view/graph/Graph.kt
@@ -1,11 +1,9 @@
-package app.view
+package app.view.graph
 
 import androidx.compose.foundation.gestures.detectDragGestures
 import androidx.compose.foundation.layout.*
 import androidx.compose.material3.*
-import androidx.compose.material3.ButtonDefaults.textButtonColors
 import androidx.compose.runtime.*
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.geometry.Offset
@@ -13,8 +11,8 @@ import androidx.compose.ui.input.pointer.PointerEventType
 import androidx.compose.ui.input.pointer.onPointerEvent
 import androidx.compose.ui.input.pointer.pointerInput
 import androidx.compose.ui.unit.Dp
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.zIndex
+import app.view.ScreenDrag
+import app.view.ScreenScale
 import app.view.model.Node
 import kotlin.math.max
 import kotlin.math.min
@@ -22,9 +20,8 @@ import kotlin.math.min
 
 @OptIn(ExperimentalComposeUiApi::class)
 @Composable
-fun Graph(root: Node, nodeSize: Dp) {
-    val screenDrag = remember { ScreenDrag(0f, 0f) }
-    val screenScale = remember { ScreenScale(1f, Offset(0f, 0f)) }
+fun Graph(root: Node, nodeSize: Dp, screenDrag: ScreenDrag, screenScale: ScreenScale) {
+
 
     Box(modifier = Modifier.fillMaxSize().onPointerEvent(PointerEventType.Scroll) {
         screenDrag.x -= it.changes.first().scrollDelta.x * 50 // let's scroll faster brbrbr
@@ -53,15 +50,5 @@ fun Graph(root: Node, nodeSize: Dp) {
         Tree(
             rootNode = root, screenDrag = screenDrag, screenScale = screenScale, nodeSize = nodeSize
         )
-        TextButton(
-            modifier = Modifier.align(Alignment.TopEnd).zIndex(3f).padding(5.dp), onClick = {
-                screenScale.scale = 1f
-                screenScale.posRelXYScale = Offset(0f, 0f)
-                screenDrag.x = 0f
-                screenDrag.y = 0f
-            }, colors = textButtonColors(
-                containerColor = MaterialTheme.colorScheme.surface
-            )
-        ) { Text("Reset view", color = MaterialTheme.colorScheme.primary) }
     }
 }
